requirements.txt
# Core dependencies
Flask==2.3.3
mysql-connector-python==8.1.0
python-dotenv==1.0.0

# API integration
requests==2.31.0

# Security
Werkzeug==2.3.7
PyJWT==2.8.0

# Testing
pytest==7.4.0
pytest-flask==1.2.0

# Development
flake8==6.1.0
black==23.7.0 